requirements.txt
google-cloud-aiplatform==1.69.*

fastapi[standard]==0.115.*
python-multipart==0.0.12
pillow==10.4.*
dataclasses-json==0.6.*